add edit functionality, sort some components
