+++ frontend/app/(settings)/profile/registrations/page.tsx
@@ -19,7 +19,7 @@ import { getClient } from "@/lib/graphql";
 import { useUser } from "@/components/providers";
 import { Dialog } from "@/components/ui/dialog";
 import PlannerItem from "@/components/planner-item";
-import EventDialog from "@/components/event-dialog/event-dialog";
+import EventDialog from "@/components/dialog/events/event-dialog";
 import { defaultEvent } from "@/types/defaults";
 
 import {
