+++ frontend/components/dialog/events/edit-event-view.tsx
@@ -24,7 +24,7 @@ import {
   DialogFooter,
   DialogContent,
 } from "@/components/ui/dialog";
-import { useRefetch, useUser } from "../providers";
+import { useRefetch, useUser } from "../../providers";
 import { getClient } from "@/lib/graphql";
 import TextareaAutosize from "react-textarea-autosize";
 import { z } from "zod";
@@ -40,12 +40,12 @@ import {
 } from "@/components/ui/form";
 import { extractId } from "@/lib/utils";
 import { TutorialsTable } from "./tutorials-table";
-import { Input } from "../ui/input";
+import { Input } from "../../ui/input";
 import { defaultEvent } from "@/types/defaults";
-import { Switch } from "../ui/switch";
-import { BadgePicker } from "../badge-picker";
-import { DatePicker } from "../date-picker";
-import { Button } from "../ui/button";
+import { Switch } from "../../ui/switch";
+import { BadgePicker } from "../../badge-picker";
+import { DatePicker } from "../../date-picker";
+import { Button } from "../../ui/button";
 import { usePathname } from "next/navigation";
 import { toast } from "sonner";
