+++ frontend/components/dialog/umbrellas/umbrella-dialog.tsx
@@ -2,9 +2,9 @@ import {Dialog, DialogContent, DialogHeader, DialogTitle} from "@/components/ui/
 import React from "react";
 import {Sprout} from "lucide-react";
 import {Event} from "@/lib/gql/generated/graphql"
-import EditUmbrellaForm from "@/app/(settings)/admin/umbrellas/edit-umbrella-form";
+import EditUmbrellaForm from "@/components/dialog/umbrellas/edit-umbrella-form";
 import {Tabs, TabsContent, TabsList, TabsTrigger} from "@/components/ui/tabs";
-import CopyUmbrellaForm from "@/app/(settings)/admin/umbrellas/copy-umbrella-form";
+import CopyUmbrellaForm from "@/components/dialog/umbrellas/copy-umbrella-form";
 
 interface EditRoomDialogProps {
   umbrella: Event;
